+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+### Added
+
+- Add `rcon exec` command to execute a Factorio console command via RCon (#87)
+- Add `rcon eval` command to evaluate a Lua script in a running Factorio server via RCon (#87)
+
 ## [0.12.0] - 2026-04-21
 
 ### Added